backend/requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
playwright
beautifulsoup4
google-generativeai
python-dotenv
uvicorn
pydantic
requests
passlib[bcrypt]
pyjwt
python-multipart
sse-starlette
